gloss training utils

+++ src/preprocess.py
@@ -102,20 +102,23 @@ class Preprocess(object):
         pickle.dump(c2idx, open(os.path.join(data_dir, 'c2idx.pkl'), 'wb'))
 
         # making key
-        key = set([])
-        v2idx = pickle.load(open(os.path.join(data_dir, 'corpus' + '.v2idx.pkl'), 'rb'))
-        gv2idx = pickle.load(open(os.path.join(data_dir, 'gloss' + '.v2idx.pkl'), 'rb'))
-        c_file = open(corpus_file, 'r', encoding='utf-8').readlines()
-        g_file = open(gloss_file, 'r', encoding='utf-8').readlines()
-        for c_line, g_line in zip(c_file, g_file):
-            c_words = c_line.strip().split()
-            g_words = g_line.strip().split()
-            assert len(c_words) == len(g_words)
-            for c, g in zip(c_words, g_words):
-                if c in v2idx and g in gv2idx:
-                    key.add((gv2idx[g], v2idx[c]))
+        if os.path.exists(os.path.join(data_dir, 'gloss' + '.v2idx.pkl')):
+            key = set([])
+            v2idx = pickle.load(open(os.path.join(data_dir, 'corpus' + '.v2idx.pkl'), 'rb'))
+            gv2idx = pickle.load(open(os.path.join(data_dir, 'gloss' + '.v2idx.pkl'), 'rb'))
+            c_file = open(corpus_file, 'r', encoding='utf-8').readlines()
+            g_file = open(gloss_file, 'r', encoding='utf-8').readlines()
+            for c_line, g_line in zip(c_file, g_file):
+                c_words = c_line.strip().split()
+                g_words = g_line.strip().split()
+                assert len(c_words) == len(g_words)
+                for c, g in zip(c_words, g_words):
+                    if c in v2idx and g in gv2idx:
+                        key.add((gv2idx[g], v2idx[c]))
 
-        pickle.dump(key, open(os.path.join(data_dir, 'key.pkl'), 'wb'))
+            pickle.dump(key, open(os.path.join(data_dir, 'key.pkl'), 'wb'))
+        else:
+            print('Can not make "key.pkl" file, no gloss found')
 
 
 if __name__ == '__main__':
